x_AdvancedProjects/MCP-Server-with-Authentication-and-Testing-with-Inspector/pyproject.toml
[project]
name = "mcp-server-with-authentication-and-testing-with-inspector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.10.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
]
